refactor(react-native): bypass proxy once client is initialized
