+++ packages/platforms/react-native/tests/create-client.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
 import {
   InMemoryDelivery,
   IncrementingClock,
@@ -143,42 +142,6 @@ describe('createReactNativeClient', () => {
 
       // @ts-expect-error startNavigationSpan should not exist on the client type
       expect(client.attach).toBeUndefined()
-    })
-  })
-
-  describe('singleton behavior', () => {
-    beforeEach(() => {
-      // Reset modules to clear the singleton
-      jest.resetModules()
-    })
-
-    it('should allow custom clients to be set and used as default', () => {
-      const { registerClient } = require('../lib/client')
-      const { createReactNativeClient } = require('../lib/create-client')
-      const clock = new IncrementingClock()
-
-      // Import the default client
-      const defaultClient = require('../lib/client').default
-
-      // Create and register a custom client
-      const testClient = createReactNativeClient({ clock })
-      registerClient(testClient)
-
-      // The default should proxy to the test client
-      expect(defaultClient.start).toBe(testClient.start)
-      expect(defaultClient.startSpan).toBe(testClient.startSpan)
-    })
-
-    it('should create default client if none is set', () => {
-      const createReactNativeClient = jest.spyOn(require('../lib/create-client'), 'createReactNativeClient')
-      const client = require('../lib/client')
-      const defaultClient = client.default
-
-      // createReactNativeClient should be called on first access
-      expect(createReactNativeClient).not.toHaveBeenCalled()
-      expect(defaultClient.start).toBeDefined()
-      expect(defaultClient.startSpan).toBeDefined()
-      expect(createReactNativeClient).toHaveBeenCalledTimes(1)
     })
   })
 })
